backend/ghibliz/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudinary==1.43.0
colorama==0.4.6
cryptography==44.0.2
distro==1.9.0
dj-rest-auth==7.0.1
Django==5.1.7
django-allauth==65.6.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
openai==1.69.0
pillow==11.1.0
pycparser==2.22
pydantic==2.11.0
pydantic_core==2.33.0
PyJWT==2.9.0
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stripe==11.6.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
